chore: Table component v2
