+++ lib/moon/components/text_input/subcomponents/text_input_basic.ex
@@ -25,7 +25,10 @@ defmodule Moon.Components.TextInput.TextInputBasic do
     <Container {=@disabled} {=@size} {=@class}>
       <label
         :if={@label}
-        class={"block text-moon-16 text-bulma pb-2 pointer-events-none", "opacity-30 cursor-not-allowed": @disabled}
+        class={
+          "block text-moon-16 text-bulma pb-2 pointer-events-none",
+          "opacity-30 cursor-not-allowed": @disabled
+        }
       >
         {@label}
       </label>
